Change IR features
IR features are now:
* scaled by the stats of the very first state
* contain not only mean value of their proportions, but also min, max
  and variance
* scaled in two ways - by overall number of instructions and the
  number of this particular type of instruction
* besides the current state of IR also include the difference with
  initial state

+++ lib/Wazuhl/ExperienceReplay.cpp
@@ -128,7 +128,7 @@ void ExperienceReplayImpl::createCollection(mongocxx::collection &collection,
                      << "size"
                      << 4294967296 // capped collection should have 'size'
                                    // attribute we limit it to 4GB
-                     << "max" << config::ExperienceSize << finalize;
+                     << "max" << (int)config::ExperienceSize << finalize;
       // TODO: move constants to config
 
       // collection is created by a 'document' command and not by
@@ -188,7 +188,7 @@ RecalledExperience ExperienceReplayImpl::replay() {
 
   // Randomly choose MinibatchSize number of experience entries
   auto SampleQuery = pipeline{};
-  SampleQuery.sample(MinimalSizeForDB);
+  SampleQuery.sample((int)MinimalSizeForDB);
 
   auto Cursor = ApprovedRecords.aggregate(SampleQuery);
